add electronic product and review

+++ src/main/kotlin/com/alonso/demokotlin/exception/ProductNotFoundByIdException.kt
@@ -1,5 +1,5 @@
 package com.alonso.demokotlin.exception
 
-import java.math.BigInteger
-
-class ProductNotFoundByIdException(id: BigInteger) : RuntimeException("Product with id $id not found")
+class ProductNotFoundByIdException(
+    id: Long,
+) : RuntimeException("Product with id $id not found")
